==> inst/extdata/react1_round_19/round19_twin_peaks_scripts_and_tabs/Scripts/13_draft_check_cases_in_lineage.R <==
n_negatives=rep(NA,19)
names(n_negatives)=1:19
for (i in 8:19){
lineage=readRDS(paste0("E:/Group/saved_objects/rep",i,"_lineage.rds"))
print(table(lineage$estbinres))
n_negatives[i]=sum(lineage$estbinres==0)
}
